refactor(auth): split auth profile manager
